refactor: 컴퓨터 및 유저 패키징
