+++ src/main/java/baseball/domain/Gamer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import java.util.Arrays;
 import java.util.HashSet;
