+++ src/test/java/baseball/domain/ComputerTest.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.domain;
 
+import baseball.domain.Computer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
